requirements.txt
Django==1.4.6
resolver==0.2.1
selector==0.8.11
Paste==1.7.2
PasteDeploy==1.3.3
PasteScript==1.7.3
cdap==1.0
httplib2==0.7.7
restclient==0.11.0
feedparser==4.1
Markdown==2.1.1
pyenchant==1.5.1
simplejson==3.3.0
smartypants==1.6.0.2
uuid==1.30
psycopg2==2.5.1
Pillow==1.7.8
nose==1.1.2
versiontools==1.9.1
statsd==0.5.1
pep8==1.4
pyflakes==0.5.0
flake8==1.7.0
lxml==3.2.3
fuzzywuzzy==0.1
sure==1.1.7
ipython==0.13.1
ipdb==0.7
lettuce==0.2.14
rdflib==3.1.0
selenium==2.20.0
coverage==3.5.3
requests==0.13.3
logilab-common==0.59.0
logilab-astng==0.24.1
pylint==0.26.0
six==1.2.0
python-dateutil==2.1
mimeparse==0.1.3
django-tastypie==0.9.11
factory_boy==2.1.1

djangowind==0.10.3
sorl==3.1
tagging==0.3-pre
typogrify==1.0
django-staticmedia==0.2
django-indexer==0.2.1
django-templatetag-sugar==0.1
django-paging==0.2.2
django-munin==0.1.1
South==0.8.1
django-annoying==0.7.6
django-nose==0.1.3
django-appconf==0.4.1
BeautifulSoup==3.2.1
django_compressor==1.3
django-statsd-mozilla==0.3.3
raven==0.7.0
django-bootstrap-form==2.0.3
django-debug-toolbar==0.9.4
django-waffle==0.9.0
django-jenkins==0.14.0
django-smoketest==0.1.3
django-extensions==1.1.0
django-treebeard==1.61
django-pagetree==0.12.1
django-pageblocks==0.5.9
django-quizblock==0.6.9.1
django-pagetree_export==0.4
django-stagingcontext==0.1.0
django-registration==0.8
